% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{copy_team_grades}
\alias{copy_team_grades}
\title{Copy grades from one team member to all others}
\usage{
copy_team_grades(df, roster)
}
\arguments{
\item{df}{Assessment with one graded team member}

\item{roster}{Course roster data frame}
}
\description{
Copy grades from one team member to all others
}
