% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{get_assessment}
\alias{get_assessment}
\title{Get assessment data frame}
\usage{
get_assessment(pars)
}
\arguments{
\item{pars}{List of parameters defining assignment}
}
\description{
Get assessment data frame
}
