% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{get_assignment_grades}
\alias{get_assignment_grades}
\title{Get grades for an assignments}
\usage{
get_assignment_grades(pars, roster)
}
\arguments{
\item{pars}{List of parameters defining assignment}

\item{roster}{Course roster data frame}
}
\description{
Get grades for an assignments
}
