% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_folder_sizes}
\alias{get_folder_sizes}
\title{Return data frame of folder sizes}
\usage{
get_folder_sizes(pars, roster)
}
\arguments{
\item{pars}{List of features defining an assignment}

\item{roster}{Class roster (data frame)}
}
\description{
Return data frame of folder sizes
}
