% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{get_gradebook}
\alias{get_gradebook}
\title{Get assessment data frame from gradebook}
\usage{
get_gradebook(pars)
}
\arguments{
\item{pars}{List of parameters defining assignment}
}
\description{
Get assessment data frame from gradebook
}
