% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{get_gradebook_assessment}
\alias{get_gradebook_assessment}
\title{Get assessment data frame from gradebook}
\usage{
get_gradebook_assessment(pars, url)
}
\arguments{
\item{pars}{List of parameters defining assignment}

\item{url}{url to google sheet}
}
\description{
Get assessment data frame from gradebook
}
