% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{get_local_assessment}
\alias{get_local_assessment}
\title{Get saved assessment data frame}
\usage{
get_local_assessment(pars)
}
\arguments{
\item{pars}{List of parameters defining assignment}
}
\description{
Get saved assessment data frame
}
