% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{grade_participation}
\alias{grade_participation}
\title{Update participation grade}
\usage{
grade_participation(url, roster, buffer = 1)
}
\arguments{
\item{url}{url to google sheet gradebook.}

\item{roster}{Course roster data frame.}

\item{buffer}{Number of days as a buffer. Defaults to \code{1}}
}
\description{
Update participation grade
}
