% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessment.R
\name{make_blank_assessment}
\alias{make_blank_assessment}
\title{Make blank assessment.csv file}
\usage{
make_blank_assessment(pars, roster)
}
\arguments{
\item{pars}{Parameters defining assignment}

\item{roster}{Course roster data frame}
}
\description{
Make blank assessment.csv file
}
