% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{save_assignment_grades}
\alias{save_assignment_grades}
\title{Save all grades to file}
\usage{
save_assignment_grades(pars, roster)
}
\arguments{
\item{pars}{Parameters defining assignment}

\item{roster}{Course roster data frame}
}
\description{
Save all grades to file
}
