% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{unzip_submissions}
\alias{unzip_submissions}
\title{Unzip or copy submissions}
\usage{
unzip_submissions(pars, roster, junkpathsSetting = FALSE)
}
\arguments{
\item{pars}{List of features defining an assignment}

\item{roster}{Class roster (data frame)}

\item{junkpathsSetting}{Ignore the file structure in the zipped file?
Defaults to \code{FALSE}.}
}
\description{
If submission is a zip file, it unzips the contents, otherwise it copies
the file over to \code{path}
}
