% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{get_all_grades}
\alias{get_all_grades}
\title{Get grades for all assignments}
\usage{
get_all_grades(assignments, roster)
}
\arguments{
\item{assignments}{Data frame of all assignments}

\item{roster}{Course roster data frame}

\item{amg}{Include an Alternative Minimum Grade (AMG) grade?}
}
\description{
Get grades for all assignments
}
