% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{make_reports}
\alias{make_reports}
\title{Make pdf reports of assignment feedback}
\usage{
make_reports(pars, roster, template, indiv = NULL)
}
\arguments{
\item{pars}{Parameters defining assignment}

\item{roster}{Course roster data frame}

\item{template}{qmd template for making report.}

\item{indiv}{One specific id (to build just one report instead of all)
Defaults to \code{NULL} in which case reports are made for all students.}
}
\description{
Make pdf reports of assignment feedback
}
