% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{save_final_grades}
\alias{save_final_grades}
\title{Save final grades to csv}
\usage{
save_final_grades(assignments, roster, drop = NULL, file = "grades.csv")
}
\arguments{
\item{assignments}{Data frame of all assignments}

\item{roster}{Course roster data frame}

\item{drop}{Which assignments to drop from grade computation. Should be a
named vector defining the category and number to drop.}

\item{file}{File name where to save grades as csv}
}
\description{
Saves final grades to a csv file located at \code{file}
}
