% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feedback.R
\name{update_feedback}
\alias{update_feedback}
\title{Update feedback for all assignments}
\usage{
update_feedback(assignments, roster, path_box, indiv = NULL)
}
\arguments{
\item{assignments}{Assignments data frame}

\item{roster}{Course roster data frame}

\item{path_box}{Path to root box folder}

\item{indiv}{One specific id (to build just one report instead of all)
Defaults to \code{NULL} in which case reports are made for all students.}
}
\description{
Update feedback for all assignments in Box folders
}
