% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{update_grades}
\alias{update_grades}
\title{Update grades on Box}
\usage{
update_grades(assignments, roster, path_box)
}
\arguments{
\item{assignments}{Assignments data frame}

\item{roster}{Course roster data frame}

\item{path_box}{Path to root box folder}
}
\description{
Update grades on Box
}
